relocated controller into controller folder, fixed link to controller in rotuer
