+++ app/controllers/challenges_controller.rb
